refactor(chain): split context graph reverse resolver
